Store response body and status in Almodovar::HttpError
This is useful when there is an error in a call to the Movida API, for example,
and you want to be able to access the actual error message in the response 
body, to know why the call failed. 
For example, in the Movida API if you try to create a record and some 
validation fails the response xml will contain the validation error message. 
Having access to this makes debugging much easier.